% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpbart_call.R
\name{mpbart_call}
\alias{mpbart_call}
\title{Multinomial Probit Bayesian Additive Regression Trees}
\usage{
mpbart_call(formula, data, base, test.data = NULL, Prior = NULL,
  Mcmc = NULL)
}
\description{
A function to implement multinomial probit regression via Bayesian Addition Regression Trees using partial marginal data augmentation.
}

