% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpbart.r
\name{rmpbart}
\alias{rmpbart}
\title{Multinomial Probit Bayesian Additive Regression Trees}
\usage{
rmpbart(Data, testData, Prior, Mcmc, seedvalue)
}
\arguments{
\item{Data}{Training data list that includes the number of choices/classes p, the response y and predictors X: list(p, y, X)}

\item{testData}{Testing data list that includes predictors denoted by X: list(X)}

\item{Prior}{List of Priors for MPBART: e.g., Prior = list(nu=p+2,  V= diag(p - 1), ntrees=200,  kfac=2.0,  pbd=1.0, pb=0.5 , beta = 2.0, alpha = 0.95, nc = 100, priorindep = 0,  minobsnode = 10)}

\item{Mcmc}{List of MCMC starting values, burn-in ...: e.g.,     list(sigma0 = diag(p - 1), keep = 1, burn = 100, ndraws = 1000, keep_sigma_draws=FALSE)}

\item{seedvalue}{random seed value: e.g., seedvalue = 99}
}
\description{
A function to implement multinomial probit regression via Bayesian Addition Regression Trees using partial marginal data augmentation.
}

